ensure last ball is always stable in view of parents, workaround past validation error

+++ network.js
@@ -817,6 +817,10 @@ function purgeJunkUnhandledJoints(){
 }
 
 function purgeJointAndDependenciesAndNotifyPeers(objJoint, error, onDone){
+	if (error.indexOf('is not stable in view of your parents') >= 0){ // give it a chance to be retried after adding other units
+		eventBus.emit('nonfatal_error', "error on unit "+objJoint.unit.unit+": "+error+"; "+JSON.stringify(objJoint), new Error());
+		return onDone();
+	}
 	joint_storage.purgeJointAndDependencies(
 		objJoint, 
 		error,
